[KAIZEN-0] flytte inParallel utility til felles sted
