+++ web/src/main/java/no/nav/modiapersonoversikt/service/saker/SakerServiceImpl.kt
@@ -1,8 +1,5 @@
 package no.nav.modiapersonoversikt.service.saker
 
-import no.nav.common.auth.subject.SubjectHandler
-import no.nav.common.log.MDCConstants
-import no.nav.modiapersonoversikt.infrastructure.http.getCallId
 import no.nav.modiapersonoversikt.legacy.api.domain.bidragsak.generated.apis.BidragSakControllerApi
 import no.nav.modiapersonoversikt.legacy.api.domain.saker.Sak
 import no.nav.modiapersonoversikt.legacy.api.exceptions.JournalforingFeilet
@@ -15,14 +12,11 @@ import no.nav.modiapersonoversikt.legacy.api.utils.SakerUtils
 import no.nav.modiapersonoversikt.service.saker.kilder.*
 import no.nav.modiapersonoversikt.service.saker.mediation.SakApiGateway
 import no.nav.modiapersonoversikt.service.unleash.UnleashService
+import no.nav.modiapersonoversikt.utils.ConcurrencyUtils.inParallel
 import no.nav.tjeneste.domene.brukerdialog.henvendelse.v1.behandlehenvendelse.BehandleHenvendelsePortType
 import no.nav.virksomhet.tjenester.sak.arbeidogaktivitet.v1.ArbeidOgAktivitet
 import org.slf4j.LoggerFactory
-import org.slf4j.MDC
 import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.web.context.request.RequestAttributes
-import org.springframework.web.context.request.RequestContextHolder
-import java.util.concurrent.CompletableFuture
 import java.util.function.Predicate.not
 import javax.annotation.PostConstruct
 import kotlin.contracts.ExperimentalContracts
@@ -173,15 +167,6 @@ class SakerServiceImpl : SakerService {
             )
         }
 
-        private fun <S, T> inParallel(first: () -> S, second: () -> T): Pair<S, T> {
-            val firstTask = CompletableFuture.supplyAsync(copyAuthAndMDC(first))
-            val secondTask = CompletableFuture.supplyAsync(copyAuthAndMDC(second))
-
-            CompletableFuture.allOf(firstTask, secondTask).get()
-
-            return Pair(firstTask.get(), secondTask.get())
-        }
-
         private fun sakFinnesIkkeIPsakOgGsak(sak: Sak): Boolean {
             return !(sak.finnesIPsak || sak.finnesIGsak)
         }
@@ -201,31 +186,3 @@ class SakerServiceImpl : SakerService {
 }
 
 private infix fun <T> ((T) -> Boolean).or(other: (T) -> Boolean): (T) -> Boolean = { this(it) || other(it) }
-internal fun <T> copyAuthAndMDC(fn: () -> T): () -> T {
-    val callId = getCallId()
-    val subject = SubjectHandler.getSubject()
-    val requestAttributes: RequestAttributes? = RequestContextHolder.getRequestAttributes()
-    return {
-        withRequestAttributes(requestAttributes) {
-            withCallId(callId) {
-                SubjectHandler.withSubject(subject.get(), fn)
-            }
-        }
-    }
-}
-
-fun <T> withCallId(callId: String, fn: () -> T): T {
-    val originalCallId = getCallId()
-    MDC.put(MDCConstants.MDC_CALL_ID, callId)
-    val result = fn()
-    MDC.put(MDCConstants.MDC_CALL_ID, originalCallId)
-    return result
-}
-
-fun <T> withRequestAttributes(requestAttributes: RequestAttributes?, fn: () -> T): T {
-    val original: RequestAttributes? = RequestContextHolder.getRequestAttributes()
-    RequestContextHolder.setRequestAttributes(requestAttributes)
-    val result = fn()
-    RequestContextHolder.setRequestAttributes(original)
-    return result
-}
